Fixes +enhancements
CMake full impl for the lib now completely works on Windows (tested with
VS2010 only for now but imports and build  great on VS2013)
- Fixed Crash in RtAudio when usb or firewre audio asio not detected yet
and setup was refering to previous asio, now displays smoothly other
choices in audio prefs without crashing.
- Fixed buffer sizes of 96 and 192 were not handled by RtAudio (commonly
used on interfaces such as MOTU and M-Audio)
- Fixed preprocessing issues in MSVC
- Fixed non automatically generated preprocessor definitions were needed
to add to cmake to make it work on MSVC
- Fixed non optimized compilation mode was selected in cmake builds,
runtime library used are now MTD / MT models, now also using SSE2 by
default and modern extensions of  WS2_32 lib when using jnet per
default.
- Fixed latest changes on MessageBox (finished TODO impl of return
codes) were not submitted in latest submission.
- Also lot of char str safe manip changes were not submitted in latst
submission for some obscure reasons, now should be fixed.
- Smarter OS_WIN setting handling, permitting to be fully compatible
with previous setting methods while permitting to use modern cmake
project files autogen on windows more smoothly.
- Enhanced jnet internall error handling , now permitting much easier
debugging with output console dumps echoed when connection errors
happen.
- improved compressing & archiving scripts, added an sh tarball script
for unix and also  win32 git bash users.
- other minor code warning cleanup

+++ WDL/lice/lice_texgen.cpp
@@ -330,7 +330,7 @@ void LICE_TexGen_CircNoise(LICE_IBitmap *dest, RECT *rect, float rv, float gv, f
         float yValue = ((float)i - h / 2) / h;
 
         float distValue = sqrt(xValue * xValue + yValue * yValue) + turbPower * turbulence(j, i, turbSize) / 256.0f;
-        float sineValue = 256.0f * fabs(sin(2 * xyPeriod * distValue * 3.14159));
+        float sineValue = (float) (256.0f * fabs(sin(2 * xyPeriod * distValue * 3.14159f)));
 
         float col = sineValue;
         p[j] = LICE_RGBA((int)(col*rv),(int)(col*bv),(int)(col*gv),255);
